+++ NEWS.md
@@ -11,6 +11,10 @@ editor_options:
 -   Add parameter `keep.these` to `stat_dens2d_labels()`, `stat_dens2d_filter()`, 
     `stat_dens2d_filter_g()`, `stat_dens1d_labels()`, `stat_dens1d_filter()` and 
     `stat_dens1d_filter_g()`.
+-   Update `stat_dens1d_labels()`, `stat_dens1d_filter()` and 
+    `stat_dens1d_filter_g()` to allow separate control of kept observations or
+    labels in each tail (or two regions) of the density distribution, still
+    using whole data set estimate of the local density.
 
 # ggpp 0.5.0